allow linked transactions with payments
My app seems to not actually be getting the invoices paid, so I'm not
sure exactly what's up there.  It's possible they're just not rejecting
the bad xml, or the QBO system has a bug, or I'm missing part of the
magical incantation that actually makes the payment pay the invoice.